enable nav highlighting - redirect from index to general - use NavLink's isActive feature
